render.js の依存性の解消のつづき
